requirements.txt
numpy==2.3.5
pandas==2.3.3
python_dateutil==2.9.0.post0
scipy==1.16.3
streamlit==1.52.2
